+++ examples/raw/gyroscope.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """
-:mod:`accelerometer`
-==================
+:mod:`gyroscope`
+===============
 
 Created by hbldh <henrik.blidh@nedomkull.com>
-Created on 2016-04-10
+Created on 2016-04-26
 
 """
 
@@ -21,6 +21,8 @@ from pymetawear.client import discover_devices, MetaWearClient, libmetawear
 from pymetawear.exceptions import PyMetaWearException
 from pymetawear.mbientlab.metawear.core import CartesianFloat, DataTypeId, FnDataPtr
 
+print("Warning: Incomplete example!")
+
 print("Discovering nearby MetaWear boards...")
 metawear_devices = discover_devices(timeout=2)
 if len(metawear_devices) < 1:
@@ -28,12 +30,12 @@ if len(metawear_devices) < 1:
 else:
     address = metawear_devices[0][0]
 
-c = MetaWearClient(str(address), 'pybluez', debug=True)
+c = MetaWearClient(str(address), 'pygatt', debug=True)
 print("New client created: {0}".format(c))
 
 
-def acc_callback(data):
-    """Handle a (x,y,z) accelerometer tuple."""
+def _callback(data):
+    """Handle a (x,y,z) gyroscope tuple."""
     if data.contents.type_id == DataTypeId.CARTESIAN_FLOAT:
         data_ptr = cast(data.contents.value, POINTER(CartesianFloat))
         print("X: {0}, Y: {1}, Z: {2}".format(*(data_ptr.contents.x,
@@ -42,31 +44,31 @@ def acc_callback(data):
     else:
         raise PyMetaWearException('Incorrect data type id: ' +
             str(data.contents.type_id))
-_acc_callback = FnDataPtr(acc_callback)
+the_callback = FnDataPtr(_callback)
 
 
-print("Write accelerometer settings...")
-libmetawear.mbl_mw_acc_set_odr(c.board, c_float(200.0))
-libmetawear.mbl_mw_acc_set_range(c.board, c_float(2.0))
-print("Write accelerometer config....")
-libmetawear.mbl_mw_acc_write_acceleration_config(c.board)
-print("Subscribing to accelerometer signal...")
-libmetawear.mbl_mw_acc_get_acceleration_data_signal.restype = c_long
-data_signal = c_long(libmetawear.mbl_mw_acc_get_acceleration_data_signal(c.board))
-libmetawear.mbl_mw_datasignal_subscribe(data_signal, _acc_callback)
-print("Enable acc sampling on board...")
-libmetawear.mbl_mw_acc_enable_acceleration_sampling(c.board)
-print("Start acc on board...")
-libmetawear.mbl_mw_acc_start(c.board)
+print("Set gyroscope settings...")
+libmetawear.mbl_mw_gyro_bmi160_set_odr(c.board, 9)
+libmetawear.mbl_mw_gyro_bmi160_set_range(c.board, 1)
+print("Write gyroscope config....")
+libmetawear.mbl_mw_gyro_bmi160_write_config(c.board)
+print("Subscribing to gyroscope signal...")
+libmetawear.mbl_mw_gyro_bmi160_get_rotation_data_signal.restype = c_long
+data_signal = c_long(libmetawear.mbl_mw_gyro_bmi160_get_rotation_data_signal(c.board))
+libmetawear.mbl_mw_datasignal_subscribe(data_signal, the_callback)
+print("Enable gyro sampling on board...")
+libmetawear.mbl_mw_gyro_bmi160_enable_rotation_sampling(c.board)
+print("Start gyro on board...")
+libmetawear.mbl_mw_gyro_bmi160_start(c.board)
 
 time.sleep(20.0)
 
 print("Unsubscribe to notification...")
 libmetawear.mbl_mw_datasignal_unsubscribe(data_signal)
-print("Stop accelerometer...")
-libmetawear.mbl_mw_acc_stop(c.board)
-print("Disable accelerometer sampling...")
-libmetawear.mbl_mw_acc_disable_acceleration_sampling(c.board)
+print("Stop gyroscope...")
+libmetawear.mbl_mw_gyro_bmi160_stop(c.board)
+print("Disable gyroscope sampling...")
+libmetawear.mbl_mw_gyro_bmi160_disable_rotation_sampling(c.board)
 
 time.sleep(5.0)