+++ pymetawear/modules/__init__.py
@@ -16,6 +16,7 @@ from __future__ import absolute_import
 
 from .base import PyMetaWearModule, Modules
 from .accelerometer import AccelerometerModule
+from .gyroscope import GyroscopeModule
 from .switch import SwitchModule
 from .battery import BatteryModule
 from .haptic import HapticModule